render: implement FUN_00488730 + retype DAT_0066b630 (0x00488730)
Texture-sample lookup. Adds struct TextureDesc/TextureFrame and retypes
DAT_0066b630 from unsigned int to void* (pointer at source). Logic matches;
the 32x32->64 mul + >>16 needs an inline SHRD that MSVC6 emits as __aullshr
here, so it's a near-miss on the mul-shift codegen only.

+++ src/legoland/print_sprite.c
@@ -173,7 +173,7 @@ void FUN_00485f00(struct Sprite *param_1, unsigned int param_2, unsigned int par
 // FUNCTION: LEGOLAND 0x00485f20
 void FUN_00485f20(void *ptr)
 {
-    DAT_0066b630 = (unsigned int)ptr;
+    DAT_0066b630 = ptr;
 }
 
 // FUNCTION: LEGOLAND 0x00485f30
